requirements.txt
streamlit
tensorflow
numpy
scikit-learn
matplotlib
wordcloud
PyPDF2
Pillow
